update to angular 13

+++ fe/src/app/channel/channel.component.ts
@@ -1,17 +1,17 @@
-import { NotyfFlashService } from "./../services/notyf.service";
-import { SpinnersService } from "./../spinners.service";
-import { ChannelsService } from "./../services/channels.service";
-import { UserService } from "./../services/user.service";
-import { Channel } from "../interfaces/channels.interface";
-import { Component, OnInit, Input } from "@angular/core";
+import { NotyfFlashService } from './../services/notyf.service';
+import { SpinnersService } from './../spinners.service';
+import { ChannelsService } from './../services/channels.service';
+import { UserService } from './../services/user.service';
+import { Channel } from '../interfaces/channels.interface';
+import { Component, OnInit, Input } from '@angular/core';
 
 @Component({
-  selector: "app-channel",
-  templateUrl: "./channel.component.html",
-  styleUrls: ["./channel.component.scss"],
+  selector: 'app-channel',
+  templateUrl: './channel.component.html',
+  styleUrls: ['./channel.component.scss'],
 })
 export class ChannelComponent implements OnInit {
-  @Input() channel: Channel;
+  @Input() channel?: Channel;
 
   constructor(
     private userService: UserService,
@@ -24,20 +24,26 @@ export class ChannelComponent implements OnInit {
 
   handleAddChannel() {
     if (!this.userService.user.hasPlaylist) {
-      this.userService.createUserSpotifyPlaylist().subscribe(
-        (data) => {
+      this.userService.createUserSpotifyPlaylist().subscribe({
+        next: (data) => {
+          if (!this.channel) return;
+
           this.userService.addChannelToUser(this.channel._id);
         },
-        (err) => {
+        error: (err) => {
           this.notyyFlashService.errorNotyf(err);
-        }
-      );
+        },
+      });
     } else {
+      if (!this.channel) return;
+
       this.userService.addChannelToUser(this.channel._id);
     }
   }
 
   handleRemoveChannel() {
+    if (!this.channel) return;
+
     this.userService.deleteChannelFromUser(this.channel._id);
   }
 }
